+++ lib/api/src/types.rs
@@ -76,7 +76,7 @@ impl ValFuncRef for Val {
         let export = wasmer_engine::ExportFunction {
             // TODO:
             // figure out if we ever need a value here: need testing with complicated import patterns
-            import_init_function_ptr: None,
+            metadata: None,
             vm_function: wasmer_vm::VMExportFunction {
                 address: item.func_ptr,
                 signature,