+++ lib/c-api/wasmer.h
@@ -32,11 +32,11 @@
 #define WASMER_WASI_ENABLED
 
 // This file corresponds to the following Wasmer version.
-#define WASMER_VERSION "1.0.0-beta1"
+#define WASMER_VERSION "1.0.0-beta2"
 #define WASMER_VERSION_MAJOR 1
 #define WASMER_VERSION_MINOR 0
 #define WASMER_VERSION_PATCH 0
-#define WASMER_VERSION_PRE "beta1"
+#define WASMER_VERSION_PRE "beta2"
 
 #endif // WASMER_H_PRELUDE
 
@@ -1102,12 +1102,14 @@ wasmer_result_t wasmer_instantiate(wasmer_instance_t **instance,
                                    int imports_len);
 
 /**
- * Gets the length in bytes of the last error if any.
+ * Gets the length in bytes of the last error if any, zero otherwise.
  *
  * This can be used to dynamically allocate a buffer with the correct number of
  * bytes needed to store a message.
  *
- * See `wasmer_last_error_message()` to get a full example.
+ * # Example
+ *
+ * See this module's documentation to get a complete example.
  */
 int wasmer_last_error_length(void);
 
@@ -1117,29 +1119,23 @@ int wasmer_last_error_length(void);
  *
  * The `length` parameter must be large enough to store the last
  * error message. Ideally, the value should come from
- * `wasmer_last_error_length()`.
+ * [`wasmer_last_error_length`].
  *
  * The function returns the length of the string in bytes, `-1` if an
  * error occurs. Potential errors are:
  *
- *  * The buffer is a null pointer,
- *  * The buffer is too small to hold the error message.
+ *  * The `buffer` is a null pointer,
+ *  * The `buffer` is too small to hold the error message.
  *
  * Note: The error message always has a trailing NUL character.
  *
- * Example:
+ * Important note: If the provided `buffer` is non-null, once this
+ * function has been called, regardless whether it fails or succeeds,
+ * the error is cleared.
  *
- * ```c
- * int error_length = wasmer_last_error_length();
+ * # Example
  *
- * if (error_length > 0) {
- *     char *error_message = malloc(error_length);
- *     wasmer_last_error_message(error_message, error_length);
- *     printf("Error message: `%s`\n", error_message);
- * } else {
- *     printf("No error message\n");
- * }
- * ```
+ * See this module's documentation to get a complete example.
  */
 int wasmer_last_error_message(char *buffer, int length);
 
@@ -1435,34 +1431,14 @@ bool wasmer_validate(const uint8_t *wasm_bytes, uint32_t wasm_bytes_len);
  *
  * # Example
  *
- * ```rust
- * # use inline_c::assert_c;
- * # fn main() {
- * #    (assert_c! {
- * # #include "tests/wasmer_wasm.h"
- * #
- * int main() {
- *     // Get and print the version.
- *     const char* version = wasmer_version();
- *     printf("%s", version);
- *
- *     // No need to free the string. It's statically allocated on
- *     // the Rust side.
- *
- *     return 0;
- * }
- * #    })
- * #    .success()
- * #    .stdout(env!("CARGO_PKG_VERSION"));
- * # }
- * ```
+ * See the module's documentation.
  */
 const char *wasmer_version(void);
 
 /**
  * Get the major version of the Wasmer C API.
  *
- * See `wasmer_version` to learn more.
+ * See [`wasmer_version`] to learn more.
  *
  * # Example
  *
@@ -1500,21 +1476,21 @@ uint8_t wasmer_version_major(void);
 /**
  * Get the minor version of the Wasmer C API.
  *
- * See `wasmer_version_major` to learn more and get an example.
+ * See [`wasmer_version_major`] to learn more and get an example.
  */
 uint8_t wasmer_version_minor(void);
 
 /**
  * Get the patch version of the Wasmer C API.
  *
- * See `wasmer_version_major` to learn more and get an example.
+ * See [`wasmer_version_major`] to learn more and get an example.
  */
 uint8_t wasmer_version_patch(void);
 
 /**
  * Get the minor version of the Wasmer C API.
  *
- * See `wasmer_version_major` to learn more.
+ * See [`wasmer_version_major`] to learn more.
  *
  * The returned string is statically allocated. It must _not_ be
  * freed!